+++ src/Sesion/HomeInit.js
@@ -1,12 +1,11 @@
 import { onNavigate } from '../main.js';
-export const Home = (rootElement) => {
+export const HomeInit = (rootElement) => {
   const ContainerHome = `<body id="redireccionamiento">
                             <a href="./Login"><img class="ImageLogo" src="../image/Logo.png"/></a>
                             <img class="ImageCharge" src="../image/charge.gif"/> 
                                                     
                           </body> `;
   rootElement.innerHTML = ContainerHome;
-
   // redireccionamienro
  setTimeout(Redireccionamiento, 3500);
   function Redireccionamiento() {
