+++ src/firebase/configuracionFB.js
@@ -1,0 +1,10 @@
+export const firebaseConfig = {
+    apiKey: "AIzaSyAdpQXKg55Cbra_BW6AUNq5qhv38V2hF88",
+    authDomain: "red-social-asia-mania.firebaseapp.com",
+    databaseURL: "https://red-social-asia-mania-default-rtdb.firebaseio.com",
+    projectId: "red-social-asia-mania",
+    storageBucket: "red-social-asia-mania.appspot.com",
+    messagingSenderId: "1078986064606",
+    appId: "1:1078986064606:web:4a0cb964406a768abab84f",
+    measurementId: "G-SQTD9N5WER"
+  };
