+++ src/firebase/initFB.js
@@ -1,0 +1,7 @@
+import{ initializeApp, getFirestore,getAuth } from './importFB.js';
+import { firebaseConfig } from './configuracionFB.js';
+
+export const app = initializeApp(firebaseConfig);
+export const db = getFirestore();
+export const auth = getAuth(app);
+
