+++ src/index.html
@@ -39,7 +39,7 @@
   </script>---->
 
 <script type="module" src="https://www.gstatic.com/firebasejs/9.4.0/firebase-app.js"></script>
-<script type="module" src="./firebase/conectorFB.js"></script>
+<script type="module" src="./firebase/configuracionFB.js"></script>
 <script src = " https://unpkg.com/sweetalert/dist/sweetalert.min.js " > </script> 
 <script type="module" src="main.js"></script>
 </body>
